apps/streamlit/demo/requirements.txt
streamlit==1.9.2
streamlit_option_menu==0.3.2
python-dotenv==0.20.0